Test if all items in an object evaluate to TRUE.

All() is a predicate functional that takes a predicate function .f and an iterable object .x and:

iterates over each item i in object .x,

evaluates .f(i),

and ultimately returns TRUE if all items i in object .x evaluate as TRUE.

All(.x, .f, ..., na.rm = FALSE)

Arguments

.x
An iterable object.

.f
A predicate function.

...
Further arguments passed to the predicate function.

na.rm
A logical value indicating whether NA values should be stripped before the computation proceeds.

Value

A logical value indicating if all items evaluated as TRUE.

See also

Any to test if all items in an object evaluate to TRUE.

Other predicate functionals: Any; Reject

Examples

# Examples
data(mtcars)
All(mtcars, is.numeric) # TRUE

#> [1] TRUE

All(mtcars, is.character) # FALSE

#> [1] FALSE

mtcars$am <- factor(mtcars$am)
All(mtcars, is.numeric) # FALSE

#> [1] FALSE

All(mtcars, is.factor) # FALSE

#> [1] FALSE

# Handles NAs and NULLs
All(list(NA, "3", NULL), is.numeric) # FALSE

#> [1] FALSE

All(list(NA, 3, NULL), is.numeric) # FALSE

#> [1] FALSE

All(list(NA, "3", NULL, 5), is.numeric) # FALSE

#> [1] FALSE

# Use na.rm = TRUE to remove NAs and NULLS
All(list(NA, TRUE), Identity) # NA

#> [1] NA

All(list(NA, TRUE), Identity, na.rm = TRUE) # TRUE

#> [1] TRUE